refactor: Packing Modal to compodable
